refactor: rename listall to items or Items
