+++ client/components/index.js
@@ -9,7 +9,7 @@ export {default as MapSingleItem} from './MapSingleItem'
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
-export {default as ListAll} from './listAll'
+export {default as Items} from './Items'
 export {default as temp} from './temp'
 export {default as NewItemForm} from './NewItemForm'
 export {default as SingleView} from './singleView'
