+++ client/components/mobileFooter.js
@@ -7,7 +7,7 @@ import {Link} from 'react-router-dom'
 const mobileFooter = (props) => (
   <div className="row cols-3 gx-2 bg-secondary">
     <div className="col text-center">
-      <Link to="/listall" className="text-decoration-none m-0">
+      <Link to="/items" className="text-decoration-none m-0">
         <i
           className="bi bi-house text-white"
           style={{
