+++ client/components/navbar.js
@@ -10,7 +10,7 @@ const Navbar = ({handleClick, isLoggedIn}) => (
   <>
     <nav className="navbar navbar-expand-lg navbar-light bg-light">
       <div className="container-fluid">
-        <Link to="/listall" className="text-decoration-none text-dark m-0">
+        <Link to="/items" className="text-decoration-none text-dark m-0">
           <span className="navbar-brand">
             <strong>Freeshare</strong>
           </span>
